recode: move all development tooling into `bin/dev.src`
There was quite a bit of development-related tooling (both aliases and functions)
in the "~/custom-shell". The "bin/dev.src" script was named "dev.src" because that's
where the "dev" *function* was defined, but it makes more sense to move *all* of the
dev-related stuff into it, so that's the refactor here.

Also included in this change:
- The `baudit` and `naudit` aliases have been dropped in favor of
  their (previously-defined) shorter counterparts: `ba` and `na`.
- The `k` alias defined in the "~/custom-shell" development section
  was an identical duplicate, so was not carried over to "dev.src".
- As Git tooling is not really *development* related, it's all been
  kept within "~/custom-shell" and just moved to the alias section.
- The `brubocop` alias is now a `rubocop` *function*.
- The `brubowatch` function has been renamed to `rubowatch`.
- Both rubocop helper functions now invoke Rubocop with `--server`.
- Several Node/NPM-related aliases and functions have been tweaked.