create: `rails-db`/`pending-migrations`/`pending-migrations?` helpers
This also adds a `pending-migrations?` check to `db+`,
speeding things up considerably when there are no migrations to run.

Then we're leveraging *that* speedup by running `db+` as part of
`rs`/`rs1`, taking a 0.5s penalty on running rails servers in exchange
for ensuring we never have to see the "migrations pending" error again.